// src/components/EmployeeDetails/types.ts
// Shared types for Employee Details components
export interface EditFormData {
  first_name: string;
  last_name: string;
  first_native_name: string;
  middle_native_name: string;
  last_native_name: string;
  department: string;
  building: string;
  room: string;
  desk_number: string;
  date_birth_year: string;
  date_birth_month: string;
  date_birth_day: string;
  manager_id: string;
  phone: string;
  email: string;
  skype: string;
  cnumber: string;
  citizenship: string;
  visa1_issuing_country: string;
  visa1_type: string;
  visa1_start_date: string;
  visa1_end_date: string;
  visa2_issuing_country: string;
  visa2_type: string;
  visa2_start_date: string;
  visa2_end_date: string;
}